feat: comment 컴포넌트에 댓글 삭제 기능 구현
